chore: refactor posting
