+++ packages/frontend/src/components/example/PostMessage.tsx
@@ -1,163 +1,8 @@
 import { useEffect, useState } from 'react';
-import axiosBase from 'axios';
-import {
-  NYM_CODE_TYPE,
-  NymProver,
-  eip712MsgHash,
-  DOMAIN,
-  Content,
-  AttestationScheme,
-  toTypedNymName,
-  CONTENT_MESSAGE_TYPES,
-  PrefixedHex,
-  EIP712TypedData,
-} from '@personaelabs/nymjs';
-import { MerkleProof } from '@personaelabs/spartan-ecdsa';
-import { ecrecover, fromRpcSig } from '@ethereumjs/util';
+import { NYM_CODE_TYPE, DOMAIN, PrefixedHex } from '@personaelabs/nymjs';
 import { useSignTypedData, useAccount } from 'wagmi';
-
-const axios = axiosBase.create({
-  baseURL: `/api/v1`,
-});
-
-type Member = {
-  pubkey: string;
-  path: string[];
-  indices: number[];
-};
-
-type ContentUserInput = {
-  title: string;
-  body: string;
-  parentId: PrefixedHex;
-};
-
-const submitPost = async (
-  content: Content,
-  attestation: string,
-  attestationScheme: AttestationScheme,
-) => {
-  const result = await axios.post(`/posts`, {
-    content,
-    attestation,
-    attestationScheme,
-  });
-
-  return result;
-};
-
-const getLatestGroup = async () => {
-  const { data } = await axios.get('/groups/latest?set=1');
-
-  const group: {
-    root: PrefixedHex;
-    members: Member[];
-  } = data;
-
-  return group;
-};
-
-export const postDoxed = async (doxedContentInput: ContentUserInput, signTypedDataAsync: any) => {
-  const group = await getLatestGroup();
-
-  // Construct the content to sign
-  const content: Content = {
-    venue: 'nouns',
-    title: doxedContentInput.title,
-    body: doxedContentInput.body,
-    parentId: doxedContentInput.parentId ? doxedContentInput.parentId : '0x0',
-    groupRoot: group.root,
-    timestamp: Math.round(Date.now() / 1000),
-  };
-
-  // Request the user to sign the content
-  const attestation = await signTypedDataAsync({
-    primaryType: 'Post',
-    domain: DOMAIN,
-    types: CONTENT_MESSAGE_TYPES,
-    message: content,
-  });
-
-  const result = await submitPost(content, attestation, AttestationScheme.EIP712);
-
-  console.log('Created a non-pseudonymous post! postId', result.data.postId);
-};
-
-export const postPseudo = async (
-  nymName: string,
-  nymSig: any,
-  nymContentInput: ContentUserInput,
-  signTypedDataAsync: any,
-) => {
-  if (nymName && nymSig) {
-    const group = await getLatestGroup();
-    const typedNymName = toTypedNymName(nymName);
-    const userPubKey = getPubKeyFromEIP712Sig(typedNymName, nymSig);
-
-    // Get the user's merkle proof
-    const userMerkleProof = group.members.find((member) => member.pubkey === userPubKey);
-
-    // Throw if merkle proof not found (implies user not in set)
-    if (!userMerkleProof) {
-      throw new Error('User not found in set');
-    }
-
-    const merkleProof: MerkleProof = {
-      pathIndices: userMerkleProof?.indices,
-      siblings: userMerkleProof?.path.map((sibling) => BigInt(sibling)),
-      root: BigInt(group.root),
-    };
-
-    // Construct the content to sign
-    const content: Content = {
-      venue: 'nouns',
-      title: nymContentInput.title,
-      body: nymContentInput.body,
-      parentId: nymContentInput.parentId ? nymContentInput.parentId : '0x0',
-      groupRoot: group.root,
-      timestamp: Math.round(Date.now() / 1000),
-    };
-
-    // Request the user to sign the content
-    const contentSig = await signTypedDataAsync({
-      primaryType: 'Post',
-      domain: DOMAIN,
-      types: CONTENT_MESSAGE_TYPES,
-      message: content,
-    });
-
-    // Setup the prover
-    const nymProver = new NymProver({
-      enableProfiler: true,
-    });
-
-    await nymProver.initWasm();
-
-    // Generate the proof
-    const attestation = await nymProver.prove(
-      nymName,
-      content,
-      nymSig as string,
-      contentSig as string,
-      merkleProof,
-    );
-
-    const attestationHex = Buffer.from(attestation).toString('hex');
-
-    const result = await submitPost(content, attestationHex, AttestationScheme.Nym);
-    console.log('Created a pseudonymous post! postId', result.data.postId);
-  }
-};
-
-const getPubKeyFromEIP712Sig = (typedData: EIP712TypedData, sig: string): string => {
-  const { v, r, s } = fromRpcSig(sig);
-  return `0x${ecrecover(
-    eip712MsgHash(typedData.domain, typedData.types, typedData.value),
-    v,
-    r,
-    s,
-  ).toString('hex')}`;
-};
+import { ContentUserInput } from '@/types/components';
+import { postDoxed, postPseudo } from '@/lib/actions';
 
 const ExamplePost = () => {
   const { address } = useAccount();
